# cmd/api/handler/products.go
package handler

import (
	"CaliYa/core/domain/ports"
	calierrors "CaliYa/core/errors"
	"errors"
	"net/http"

	"github.com/google/uuid"
	"github.com/labstack/echo/v4"
)

type Products interface {
	RegisterProducts(c echo.Context) error
	GetProductsByCategory(c echo.Context) error
	GetAdicionesGyCategory(c echo.Context) error
}

type products struct {
	app ports.ProductsApp
}

func NewProducts(app ports.ProductsApp) Products {
	return &products{app}
}

func (p *products) RegisterProducts(c echo.Context) error {

	ctx := c.Request().Context()

	err := p.app.RegisterProduct(ctx)
	if err != nil {
		return err
	}

	return c.JSON(http.StatusOK, "ok")
}

// GetProductByCategory godoc
// @Tags Products
// @Summary End Point para obtener todos los productos de una misma categoria.
// @Produce json
// @Param category path string true "Este es el nombre de la categoria ej:/products/category/hamburguesas"
// @Success 200 {object} []models.Items
// @Failure 404
// @Failure 500
// @Router /products/category/{category} [get]
func (p *products) GetProductsByCategory(c echo.Context) error {

	ctx := c.Request().Context()

	category := c.Param("category")

	items, err := p.app.GetProductByCategory(ctx, category)
	if err != nil {
		switch {
		case errors.Is(err, calierrors.ErrNotFound):
			echo.NewHTTPError(http.StatusNotFound, err)
		default:
			echo.NewHTTPError(http.StatusInternalServerError, err)
		}

	}

	return c.JSON(http.StatusOK, items)
}

// GetAdicionesGyCategory godoc
// @Tags Products
// @Summary End Point para obtener todos las adiciones de una misma categoria.
// @Produce json
// @Param category_id query string true "Este es el id de la categoria, debe ser formato UUID"
// @Success 200 {object} []models.Items
// @Failure 404
// @Failure 500
// @Router /products/adiciones [get]
func (p *products) GetAdicionesGyCategory(c echo.Context) error {

	ctx := c.Request().Context()

	id := c.QueryParam("category_id")

	adiciones, err := p.app.GetAditionsByCategory(ctx, uuid.MustParse(id))
	if err != nil {
		return err
	}

	return c.JSON(http.StatusOK, adiciones)
}