# docs/swagger.yaml
basePath: /api
definitions:
  dto.ShopResponse:
    properties:
      address:
        type: string
      home_phone:
        type: string
      id:
        type: string
      logo:
        type: string
      name:
        type: string
      opened:
        type: boolean
      score:
        type: number
      tag:
        type: string
      type:
        type: string
    type: object
  models.Categories:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.Items'
        type: array
      name:
        type: string
      shop_id:
        type: string
    type: object
  models.Items:
    properties:
      description:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: integer
      score:
        type: integer
      shop:
        $ref: '#/definitions/models.ProductsShops'
      shop_id:
        type: string
    type: object
  models.ProductsShops:
    properties:
      address:
        type: string
      categories:
        items:
          $ref: '#/definitions/models.Categories'
        type: array
      id:
        type: string
      logo_image:
        type: string
      name:
        type: string
      opened:
        type: boolean
      score:
        type: number
      tag:
        type: string
    type: object
info:
  contact: {}
  title: Documentation API CaliYa
  version: 1.0.1
paths:
  /products/adiciones:
    get:
      parameters:
      - description: Este es el id de la categoria, debe ser formato UUID
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Items'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: End Point para obtener todos las adiciones de una misma categoria.
      tags:
      - Products
  /products/category/{category}:
    get:
      parameters:
      - description: Este es el nombre de la categoria ej:/products/category/hamburguesas
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Items'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: End Point para obtener todos los productos de una misma categoria.
      tags:
      - Products
  /shops:
    get:
      parameters:
      - description: Este es el ID del negocio, viene en formato UUID
        in: query
        name: id
        type: string
      - description: Este es el tag del negocio .ej:dmo
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductsShops'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: End Point para obtener un negocio con todos sus products, se debe enviar
        alguno de los 2 query params requeridos.
      tags:
      - Shops
  /shops/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ShopResponse'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Para que no me fastidies de cual es la ruta
      tags:
      - Shops
schemes:
- http
swagger: "2.0"